+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to parse-xml are documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 4.1.0 (git)
+## 4.1.0 (2023-02-04)
 
 ### Added